requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Ultrasound-Nerve-Segmentation/code/preprocessing.py: 4
Pillow == 5.1.0

# Ultrasound-Nerve-Segmentation/code/dataset.py: 5,6
# Ultrasound-Nerve-Segmentation/code/training.py: 15,16
albumentations == 0.4.5

# Ultrasound-Nerve-Segmentation/code/preprocessing.py: 3
# Ultrasound-Nerve-Segmentation/code/training.py: 12
matplotlib == 2.1.1

# Ultrasound-Nerve-Segmentation/code/dataset.py: 3
# Ultrasound-Nerve-Segmentation/code/preprocessing.py: 1
# Ultrasound-Nerve-Segmentation/code/submission.py: 8
# Ultrasound-Nerve-Segmentation/code/training.py: 13
numpy == 1.13.3

# Ultrasound-Nerve-Segmentation/code/dataset.py: 4
# Ultrasound-Nerve-Segmentation/code/preprocessing.py: 2
# Ultrasound-Nerve-Segmentation/code/submission.py: 9
# Ultrasound-Nerve-Segmentation/code/training.py: 14
pandas == 0.22.0

# Ultrasound-Nerve-Segmentation/code/training.py: 4
scikit_learn == 0.22.1

# Ultrasound-Nerve-Segmentation/code/dataset.py: 2
# Ultrasound-Nerve-Segmentation/code/loss_function.py: 2,4
# Ultrasound-Nerve-Segmentation/code/models.py: 2,4
# Ultrasound-Nerve-Segmentation/code/submission.py: 4,6
# Ultrasound-Nerve-Segmentation/code/training.py: 18
torch == 1.4.0+cpu

# Ultrasound-Nerve-Segmentation/code/dataset.py: 1
# Ultrasound-Nerve-Segmentation/code/loss_function.py: 3
# Ultrasound-Nerve-Segmentation/code/models.py: 3
# Ultrasound-Nerve-Segmentation/code/submission.py: 5
# Ultrasound-Nerve-Segmentation/code/training.py: 17
torchvision == 0.5.0+cpu
